feature (comments) view post's comment
checking if a post exist
check if a post is published
view comment

[Delivers ##163682238]
