+++ server/src/controllers/index.js
@@ -1,6 +1,5 @@
 import authController from './auth';
 import userController from './users';
 import postController from './posts';
-import commentController from './comments';
 
-export { authController, postController, userController, commentController };
+export { authController, postController, userController };
